feat: support GitHub repositories in scan
